Do not trim first characters of messages in error replies.
In the past Redis returned only "-ERR" for any kind of error occurred
on the server using the error reply type, so the "ERR" prefix seemed
kind of superfluous. With recent versions, Redis started using this
prefix to indicate the actual kind of error (e.g. operations on wrong
types return "-WRONGTYPE").