Implement phpiredis_utils_crc16().
This new utility function is useful to implement redis-cluster on top
of phpiredis since crc16 is at the base of the algorithm used to map
keys to hash slots. Compared to a pure-PHP implementation of CRC16 it
is ~10 times faster (PHP 7.0.7) when calculating the hash for strings
of 10 bytes. Examples of PHP projects that can benefit from it in the
immediate are nrk/predis and jarnix/LiteRedis.

phpiredis uses the reference implementation in ANSI C made available
on http://redis.io/topics/cluster-spec#appendix (compatible license).